+++ src/cljs/cxengage_javascript_sdk/modules/voice.cljs
@@ -45,7 +45,7 @@
      (send-interrupt module type (merge (iu/extract-params client-params) {:callback (first others)}))))
   ([module type client-params]
    (let [client-params (iu/extract-params client-params)
-         {:keys [callback interaction-id resource-id queue-id transfer-extension transfer-type]} client-params
+         {:keys [callback interaction-id resource-id queue-id transfer-extension transfer-resource-id  transfer-type]} client-params
          transfer-type (if (= transfer-type "cold") "cold-transfer" "warm-transfer")
          simple-interrupt-body {:resource-id (state/get-active-user-id)}
          interrupt-params (case type
@@ -92,9 +92,11 @@
                                                                      :resource-id (state/get-active-user-id)
                                                                      :transfer-type transfer-type}}
                             :cancel-transfer {:validation ::generic-voice-interaction-fn-params
-                                              :interrupt-type "cancel-transfer"
+                                              :interrupt-type "transfer-cancel"
                                               :topic (p/get-topic :cancel-transfer-acknowledged)
-                                              :interrupt-body simple-interrupt-body})]
+                                              :interrupt-body {:transfer-resource-id transfer-resource-id
+                                                               :resource-id (state/get-active-user-id)
+                                                               :transfer-type transfer-type}})]
      (if-not (s/valid? (:validation interrupt-params) client-params)
        (p/publish {:topic (:topic interrupt-params)
                    :error (e/invalid-args-error (s/explain-data (:validation interrupt-params) client-params))